@@ -5,39 +5,74 @@
   "scripts": {
     "dev": "next dev",
     "build": "next build",
+    "analyze": "ANALYZE=true next build",
     "start": "next start",
-    "lint": "next lint"
+    "typecheck": "tsc --noEmit",
+    "lint": "next lint",
+    "jest": "jest",
+    "jest:watch": "jest --watch",
+    "prettier:check": "prettier --check \"**/*.{ts,tsx}\"",
+    "prettier:write": "prettier --write \"**/*.{ts,tsx}\"",
+    "test": "npm run prettier:check && npm run lint && npm run typecheck && npm run jest",
+    "storybook": "storybook dev -p 6006",
+    "storybook:build": "storybook build"
   },
   "dependencies": {
-    "@emotion/react": "^11.11.1",
-    "@emotion/server": "^11.11.0",
     "@firebase/app": "^0.9.20",
     "@mantine/carousel": "^7.1.5",
     "@mantine/core": "^7.1.5",
     "@mantine/hooks": "^7.1.5",
     "@mantine/next": "^6.0.21",
     "@mantine/notifications": "^7.1.5",
+    "@mantine/styles": "^6.0.21",
+    "@next/bundle-analyzer": "^13.5.6",
     "@tabler/icons-react": "^2.39.0",
-    "encoding": "^0.1.13",
     "firebase": "^10.5.0",
     "next": "13.5.6",
-    "node-fetch": "^3.3.2",
     "qrcode.react": "^3.1.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-firebase-hooks": "^5.1.1",
-    "sharp": "^0.32.6",
-    "webpack": "^5.89.0"
+    "sharp": "^0.32.6"
   },
   "devDependencies": {
-    "@types/node": "^20",
-    "@types/react": "^18",
+    "@babel/core": "^7.22.1",
+    "@next/eslint-plugin-next": "^13.5.3",
+    "@storybook/addon-essentials": "^7.4.3",
+    "@storybook/addon-styling": "^1.0.8",
+    "@storybook/blocks": "^7.4.3",
+    "@storybook/nextjs": "^7.4.3",
+    "@storybook/react": "^7.4.3",
+    "@testing-library/dom": "^9.3.0",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.4.3",
+    "@types/jest": "^29.5.1",
+    "@types/node": "^20.2.5",
+    "@types/react": "18.2.7",
     "@types/react-dom": "^18",
-    "autoprefixer": "^10",
-    "eslint": "^8",
-    "eslint-config-next": "13.5.6",
-    "postcss": "^8.4.31",
-    "tailwindcss": "^3.3.4",
-    "typescript": "^5"
+    "@typescript-eslint/eslint-plugin": "^5.59.7",
+    "@typescript-eslint/parser": "^5.59.7",
+    "babel-loader": "^9.1.2",
+    "eslint": "^8.41.0",
+    "eslint-config-airbnb": "19.0.4",
+    "eslint-config-airbnb-typescript": "^17.0.0",
+    "eslint-config-mantine": "2.0.0",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-jest": "^27.2.1",
+    "eslint-plugin-jsx-a11y": "^6.7.1",
+    "eslint-plugin-react": "^7.32.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-testing-library": "^5.11.0",
+    "jest": "^29.5.0",
+    "jest-environment-jsdom": "^29.5.0",
+    "postcss": "^8.4.23",
+    "postcss-preset-mantine": "1.9.0",
+    "postcss-simple-vars": "^7.0.1",
+    "prettier": "^2.8.8",
+    "storybook": "^7.4.3",
+    "storybook-dark-mode": "^3.0.0",
+    "ts-jest": "^29.1.0",
+    "typescript": "5.2.2"
   }
 }
