@@ -1,10 +1,10 @@
-"use client"
+'use client';
 
-import { useRouter } from 'next/router';
-import React, {MouseEventHandler, useState} from 'react';
+import React, { MouseEvent, useState } from 'react';
+import { useRouter } from 'next/navigation';
 import Link from 'next/link';
 import { getAuth, signInWithEmailAndPassword } from 'firebase/auth';
-import { useAuthState } from "react-firebase-hooks/auth";
+import { useAuthState } from 'react-firebase-hooks/auth';
 import { IconCheck, IconExclamationCircle } from '@tabler/icons-react';
 import {
     Button,
@@ -15,66 +15,57 @@ import {
     Loader,
     Stack,
     TextInput,
-    Title
+    Title,
 } from '@mantine/core';
 import { notifications } from '@mantine/notifications';
-import firebase from "../../firebaseApp";
+import firebase from '../../firebaseApp';
 
 export default function DozentLogin() {
     const auth = getAuth(firebase);
+    const router = useRouter();
     const [email, setEmail] = useState('');
     const [profPasswort, setProfPasswort] = useState('');
     const [user, loading, error] = useAuthState(auth);
-    //const router = useRouter();
 
-    const handleSubmit = async (e: React.MouseEvent<HTMLButtonElement>) => {
+    const showNotification = (title: string, message: string, color: string) => {
+        notifications.update({
+            id: 'load-data',
+            loading: false,
+            color,
+            title,
+            message,
+            icon: color === 'red' ? <IconExclamationCircle size="1rem" /> : <IconCheck size="1rem" />,
+        });
+    };
+
+    const handleLogin = async (e: MouseEvent<HTMLButtonElement>) => {
         e.preventDefault();
         notifications.show({
             id: 'load-data',
-            color: "brand.0",
+            color: 'brand.0',
             loading: true,
             title: 'Versuche Login',
             message: 'Du kannst das noch nicht schließen.',
             autoClose: false,
             withCloseButton: false,
         });
+
         if (email === '' || profPasswort === '') {
-            notifications.update({
-                id: 'load-data',
-                loading: false,
-                color: 'red',
-                title: 'Upps😵‍💫',
-                message: 'Bitte geben Sie ihre Nutzerinformaitonen ein!',
-                icon: <IconExclamationCircle size="1rem" />
-            });
+            showNotification('Upps😵‍💫', 'Bitte geben Sie Ihre Nutzerinformationen ein!', 'red');
         } else {
             try {
                 await signInWithEmailAndPassword(auth, email, profPasswort);
-                notifications.update({
-                    id: 'load-data',
-                    loading: false,
-                    color: 'teal',
-                    title: 'Success',
-                    message: 'Login erfolgreich',
-                    icon: <IconCheck size="1rem" />
-                });
-                //router.push('/overview');
-            } catch (error) {
-                console.error(error);
-                notifications.update({
-                    id: 'load-data',
-                    loading: false,
-                    color: 'red',
-                    title: 'Upps, Anmeldung fehlgeschlagen. 😕',
-                    message: 'Bitte überprüfe deine Anmeldeinformationen!',
-                    icon: <IconExclamationCircle size="1rem" />
-                });
+                showNotification('Success', 'Login erfolgreich', 'teal');
+                router.push('/overview');
+            } catch (err) {
+                console.error(err);
+                showNotification('Upps, Anmeldung fehlgeschlagen. 😕', 'Bitte überprüfe deine Anmeldeinformationen!', 'red');
             }
         }
     };
 
     if (user) {
-        //router.push('/overview');
+        router.push('/overview');
     } else {
         if (loading) {
             return (
@@ -82,50 +73,45 @@ export default function DozentLogin() {
                     <Loader />
                 </Center>
             );
-        } else {
-            return (
-                <Center bg="brand.7" style={{ height: '100%' }}>
-                    <Container size={800} my={40}>
-                        <Title
-                            c="white"
-                            ta="center"
-                            style={{ fontFamily: 'Castellar, sans-serif', fontWeight: 800 }}
-                        >
-                            Game Theory
-                        </Title>
-                        <Paper miw={300} withBorder shadow="md" p={20} mt={20} radius="md">
-                            <Stack gap="sm">
-                                <TextInput
-                                    label="Email"
-                                    value={email}
-                                    onChange={(e) => setEmail(e.target.value.replace(/\s/g, ''))}
-                                />
-                                <PasswordInput
-                                    placeholder="Password"
-                                    label="Password"
-                                    value={profPasswort}
-                                    onChange={(e) => setProfPasswort(e.target.value.replace(/\s/g, ''))}
-                                />
-                                <Button onClick={handleSubmit} fullWidth my="xl">
-                                    Login
-                                </Button>
-                            </Stack>
-                            <Stack align="center" gap="xs">
-                                <Link href="/" style={{ textDecoration: 'none' }}>
-                                    Passwort vergessen?
-                                </Link>
-                                <Link href="/" style={{ textDecoration: 'none' }}>Registrieren</Link>
-                            </Stack>
-                            <Link href="/" style={{ textDecoration: 'none' }}>
-                                <Container fz={14} c={'darkgray'} ta={"right"} w={110} mr={0} p={0}>
-                                    Studenten Login
-                                    <Center><i className="fa fa-chevron-right"></i></Center>
-                                </Container>
-                            </Link>
-                        </Paper>
-                    </Container>
-                </Center>
-            );
         }
+        return (
+            <Center bg="brand.7" style={{ height: '100%' }}>
+                <Container size={800} my={40}>
+                    <Title c="white" ta="center" style={{ fontFamily: 'Castellar, sans-serif', fontWeight: 800 }}>
+                        Game Theory
+                    </Title>
+                    <Paper miw={300} withBorder shadow="md" p={20} mt={20} radius="md">
+                        <Stack gap="sm">
+                            <TextInput
+                              label="Email"
+                              value={email}
+                              onChange={(e) => setEmail(e.target.value.replace(/\s/g, ''))}
+                            />
+                            <PasswordInput
+                              placeholder="Password"
+                              label="Password"
+                              value={profPasswort}
+                              onChange={(e) => setProfPasswort(e.target.value.replace(/\s/g, ''))}
+                            />
+                            <Button onClick={handleLogin} fullWidth my="xl">
+                                Login
+                            </Button>
+                        </Stack>
+                        <Stack align="center" gap="xs">
+                            <Link href="/" style={{ textDecoration: 'none' }}>
+                                Passwort vergessen?
+                            </Link>
+                            <Link href="/" style={{ textDecoration: 'none' }}>Registrieren</Link>
+                        </Stack>
+                        <Link href="/" style={{ textDecoration: 'none' }}>
+                            <Container fz={14} c="darkgray" ta="right" w={110} mr={0} p={0}>
+                                Studenten Login
+                                <Center><i className="fa fa-chevron-right" /></Center>
+                            </Container>
+                        </Link>
+                    </Paper>
+                </Container>
+            </Center>
+        );
     }
 }
