@@ -1,5 +1,5 @@
 import React from 'react';
-import { Container, Title, Flex, StyleProp} from '@mantine/core';
+import { Container, Title, Flex } from '@mantine/core';
 import Image from 'next/image';
 
 import gameTheoryLogo from 'public/GameTheory_Header.png';
@@ -8,13 +8,13 @@ export default function StudentMenuBar() {
     return (
         <Container bg="brand.7" fluid px={40} p={10}>
             <Flex align="center">
-                <Image src={gameTheoryLogo} height={70} alt={"."}/>
+                <Image src={gameTheoryLogo} height={70} alt="." />
                 <Title
-                    c="white"
-                    ta="left"
+                  c="white"
+                  ta="left"
                     //justify="center"
-                    ff="Castellar, sans-serif"
-                    fw={700}
+                  ff="Castellar, sans-serif"
+                  fw={700}
                 >
                     Game Theory
                 </Title>
