@@ -1,28 +1,34 @@
-"use client"
+'use client';
 
-import React from 'react';
+import React, { ChangeEvent } from 'react';
 import './card.css';
 
-export default function Card(props) {
+interface CardProps {
+    id: string;
+    onChange: any;
+    side: string;
+}
+
+export default function Card(props: CardProps) {
     const { id, onChange, side } = props;
 
-    const handleInputChange = (event) => {
+    const handleInputChange = (event: ChangeEvent<HTMLInputElement>) => {
         const selectedSide = event.target.getAttribute('data-side');
         onChange(selectedSide);
     };
 
     return (
         <div>
-            <label className="selected-label">
+            <label htmlFor={`selected-item-${id}`} className="selected-label">
                 <input
-                    type="checkbox"
-                    name="selected-item"
-                    id={`selected-item-${id}`}
-                    value={id}
-                    onChange={handleInputChange}
-                    data-side={side}
+                  type="checkbox"
+                  name="selected-item"
+                  id={`selected-item-${id}`}
+                  value={id}
+                  onChange={handleInputChange}
+                  data-side={side}
                 />
-                <div className={`selected-content-${side}`}></div>
+                <div className={`selected-content-${side}`} />
             </label>
         </div>
     );
