@@ -1,18 +1,18 @@
-"use client"
+'use client';
 
 import React, { useState, useEffect } from 'react';
-import Card from './Card';
-import {Modal, Button, Container, Center, Text, Loader, Stack, Group, CheckboxProps} from '@mantine/core';
-//import { useDisclosure } from '@mantine/hooks';
+import { Modal, Button, Container, Center, Text, Loader, Stack, Group } from '@mantine/core';
+import { useDisclosure } from '@mantine/hooks';
 import { Carousel } from '@mantine/carousel';
+import Card from './Card';
 
 export default function CardSelection() {
     const [selectedCount, setSelectedCount] = useState(0);
     const [currentRound, setCurrentRound] = useState(0);
     const [numRedCards, setNumRedCards] = useState(0);
-    //const [opened, { open, close }] = useDisclosure(false);
+    const [opened, { open, close }] = useDisclosure(false);
 
-    const handleInputChangeCard = (event: Event) => {
+    const handleInputChangeCard = (event: React.ChangeEvent<HTMLInputElement>) => {
         const target = event.target as HTMLInputElement;
         const isChecked = target.checked;
         const side = target.getAttribute('data-side');
@@ -28,7 +28,7 @@ export default function CardSelection() {
         setNumRedCards((prevCount) => {
             if (isChecked && side === 'left') {
                 return prevCount + 1;
-            } else if (!isChecked && side === 'left') {
+            } if (!isChecked && side === 'left') {
                 return prevCount - 1;
             }
             return prevCount;
@@ -40,29 +40,31 @@ export default function CardSelection() {
         open();
     };
 
-    /*useEffect(() => {
-        // Fetch current round data
-        fetchCurrentRound();
-    }, []);*/
-
     const fetchCurrentRound = async () => {
         try {
             // Simulate fetching current round data
             // Replace this with your actual data fetching logic
             setCurrentRound(1); // Example value
         } catch (error) {
-            console.error("Error fetching data: ", error);
+            console.error('Error fetching data: ', error);
         }
     };
 
+    useEffect(() => {
+        // Fetch current round data
+        fetchCurrentRound();
+    }, []);
+
     return (
         <>
-            <Modal opened={false} onClose={open} centered withCloseButton={false}>
+            <Modal opened={opened} onClose={open} centered withCloseButton={false}>
                 <Stack align="center" justify="center">
-                    <Text style={{ fontFamily: 'Instrument Sans, sans-serif', fontWeight: 700 }}
-                          p={40}
-                          className="lbl-round"
-                          size={"18"}>
+                    <Text
+                      style={{ fontFamily: 'Instrument Sans, sans-serif', fontWeight: 700 }}
+                      p={40}
+                      className="lbl-round"
+                      size="18"
+                    >
                         <Group>
                             Du hast <Text c="#cc4444">{numRedCards}</Text> rote Karten abgegeben.
                         </Group>
@@ -73,10 +75,12 @@ export default function CardSelection() {
             </Modal>
             <Container fluid p={0}>
                 <Center>
-                    <Text style={{ fontFamily: 'Instrument Sans, sans-serif', fontWeight: 700 }}
-                          p={40}
-                          className="lbl-round"
-                          size={"19"}>
+                    <Text
+                      style={{ fontFamily: 'Instrument Sans, sans-serif', fontWeight: 700 }}
+                      p={40}
+                      className="lbl-round"
+                      size="19"
+                    >
                         Runde: {currentRound}
                     </Text>
                 </Center>
