@@ -1,6 +1,6 @@
-import StudentMenuBar from "./StudentMenuBar";
+import StudentMenuBar from './StudentMenuBar';
 
-export default function StudentLayout({ children }) {
+export default function StudentLayout({ children }:{ children: any }) {
     return (
         <>
             <StudentMenuBar />
