@@ -1,7 +1,9 @@
-import StudentLogin from "../page";
+import StudentLogin from '../page';
 
+// @ts-ignore
 export default function LinkGameID({ params }) {
-    const gameID = params.gameID;
+    const { gameID } = params;
+
     return (
         <StudentLogin gameIdIn={gameID} />
     );
