@@ -1,26 +1,22 @@
-"use client"
+'use client';
 
 import React, { useState, useEffect } from 'react';
-import DataCollection from "./datacollection";
 import { Button, Container, Center, ScrollArea, Stack, Text, NumberInput, Grid, Group, Modal } from '@mantine/core';
 import { useViewportSize, useDisclosure } from '@mantine/hooks';
+import DataCollection from './datacollection';
 
-export default function DozentGame({ params }) {
+interface DozentGameProps {
+    params: {
+        dozentgame: string; // Adjust the type according to your route's parameter type
+    };
+}
+
+export default function DozentGame({ params }: DozentGameProps) {
     const { height, width } = useViewportSize();
     const gameId = params.dozentgame;
-    const [redCardValue, setRedCardValue] = useState<number | string>(1)
+    const [redCardValue, setRedCardValue] = useState<number | string>(1);
     const [opened, { open, close }] = useDisclosure(false);
-    const [currentRound, setCurrentRound] = useState(0)
-
-    const handleNextRound = async () => {
-        await fetch('/api/rounds', {
-            method: 'POST',
-            body: JSON.stringify({gameID: gameId, redCardValue: redCardValue}),
-            headers: { 'Content-Type': 'application/json' }
-        });
-        fetchCurrentRound();
-        close();
-    };
+    const [currentRound, setCurrentRound] = useState(0);
 
     async function fetchCurrentRound() {
         try {
@@ -28,9 +24,19 @@ export default function DozentGame({ params }) {
             const data = await response.json();
             setCurrentRound(data.currentRound);
         } catch (error) {
-            console.error("Error fetching data: ", error);
+            console.error('Error fetching data: ', error);
         }
     }
+
+    const handleNextRound = async () => {
+        await fetch('/api/rounds', {
+            method: 'POST',
+            body: JSON.stringify({ gameID: gameId, redCardValue }),
+            headers: { 'Content-Type': 'application/json' },
+        });
+        fetchCurrentRound();
+        close();
+    };
 
     useEffect(() => {
         fetchCurrentRound();
@@ -46,13 +52,14 @@ export default function DozentGame({ params }) {
 
     return (
         <>
-            <Modal opened={opened} onClose={close} title={currentRound === 0 ? "Spiel Starten" : "Nächste Runde"}>
+            <Modal opened={opened} onClose={close} title={currentRound === 0 ? 'Spiel Starten' : 'Nächste Runde'}>
                 <Stack gap="xl">
                     <NumberInput
-                        type="text"
-                        label="Roter Kartenwert"
-                        value={redCardValue}
-                        onChange={setRedCardValue} />
+                      type="text"
+                      label="Roter Kartenwert"
+                      value={redCardValue}
+                      onChange={setRedCardValue}
+                    />
 
                     <Group align="right">
                         {currentRound === 0 ? (
@@ -73,13 +80,13 @@ export default function DozentGame({ params }) {
                     <Grid.Col span={5}>
                         <Stack justify="space-between" h={height - 200}>
                             <Group align="right" px={90}>
-                                <Text style={{ fontFamily: 'Instrument Sans, sans-serif', fontWeight: 700 }} >
+                                <Text style={{ fontFamily: 'Instrument Sans, sans-serif', fontWeight: 700 }}>
                                     Runde: {currentRound}
                                 </Text>
                             </Group>
-                            <Center >Rundenauswertung</Center>
+                            <Center>Rundenauswertung</Center>
                             <Group align="right" gap="xl">
-                                <Button variant="outline" color= "red" bg="white" >Spielabbruch</Button>
+                                <Button variant="outline" color="red" bg="white">Spielabbruch</Button>
                                 {currentRound === 0 ? (
                                     <Button bg="brand.0" onClick={open}>Spiel Starten</Button>
                                 ) : (
