@@ -1,13 +1,13 @@
-"use client"
+'use client';
 
 import React from 'react';
-import DozentMenueLeiste from './DozentMenueLeiste';
 import { redirect } from 'next/navigation';
 import { useAuthState } from 'react-firebase-hooks/auth';
 import { getAuth } from 'firebase/auth';
+import DozentMenueLeiste from './DozentMenueLeiste';
 import firebase from '../firebaseApp';
 
-export default function DozentenLayout({ children }) {
+export default function DozentenLayout({ children }: { children: any }) {
     const auth = getAuth(firebase);
     const [user, loading, error] = useAuthState(auth);
 
@@ -18,11 +18,9 @@ export default function DozentenLayout({ children }) {
                 <main>{children}</main>
             </>
         );
-    } else {
+    }
         if (loading) {
             return <p>Loading...</p>;
-        } else {
-            redirect('/');
         }
-    }
+            redirect('/');
 }
